# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "UserController_listUsers",
        "parameters": [
          {
            "name": "filterActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "사용자 목록을 조회합니다.",
        "tags": [
          "사용자 관리"
        ]
      }
    },
    "/user/{userId}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "사용자 상세 정보를 조회합니다.",
        "tags": [
          "사용자 관리"
        ]
      }
    },
    "/user/{userId}/toggle-activate": {
      "patch": {
        "operationId": "UserController_activateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "유저를 활성 상태를 토글합니다.",
        "tags": [
          "사용자 관리"
        ]
      }
    },
    "/user/{userId}/role": {
      "patch": {
        "operationId": "UserController_assignRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "admin"
            ]
          }
        ],
        "summary": "유저의 역할을 변경합니다.",
        "tags": [
          "사용자 관리"
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Manager",
    "description": "이벤트/보상 관리 시스템",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserConditionDto": {
        "type": "object",
        "properties": {
          "loginStreakDays": {
            "type": "number"
          },
          "referralCount": {
            "type": "number"
          }
        },
        "required": [
          "loginStreakDays",
          "referralCount"
        ]
      },
      "UserInventoryDto": {
        "type": "object",
        "properties": {
          "point": {
            "type": "number"
          },
          "coupons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "point",
          "coupons",
          "items"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string"
          },
          "condition": {
            "$ref": "#/components/schemas/UserConditionDto"
          },
          "inventory": {
            "$ref": "#/components/schemas/UserInventoryDto"
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UserListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "filterActive": {
            "type": "boolean"
          },
          "total": {
            "type": "number"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "list"
        ]
      },
...
    }
  }
}